Added ReGex
Created func cr_extractor which iterates through the roll list and compares every item to a camera roll regex. Then adds the matches to a mo, and then we write the camera rolls to rolls.text.

Created func called sr_curator which grabs column five of the csv, which is the paths column. We then iterated through this looking for sound_master using a regex, and the we store the index number in a new list. We then use a while loop to iterate through the the original path list and grab the last three indexes which are the sound rolls. These are then appended to another list, we then append the sound rolls to rolls.text.

General polishing will need to be done to the script.